Move utils/ to common/
