+++ components/carte-etat/simpop-tableur-infos-deciles.tsx
@@ -7,7 +7,7 @@ import TableRow from "@material-ui/core/TableRow";
 
 // eslint-disable-next-line no-unused-vars
 import { RootState } from "../../redux/reducers";
-import { formatNumber } from "../utils";
+import { formatNumber } from "../common";
 import styles2 from "./simpop-tableur-infos-deciles.module.scss";
 
 const styles = {
