+++ components/common/articles-inputs/parameter/Parameter.tsx
@@ -2,7 +2,7 @@ import classNames from "classnames";
 import { PureComponent } from "react";
 
 import { PlfTooltip, ReformTooltip } from "../../tooltips";
-import { formatNumber } from "../../../utils";
+import { formatNumber } from "../../utils";
 import { NumberInput } from "./number-input";
 import styles from "./Parameter.module.scss";
