+++ components/common/articles-inputs/parameter/number-input/NumberInput.tsx
@@ -1,6 +1,6 @@
 import { PureComponent } from "react";
 
-import { formatNumber } from "../../../../utils";
+import { formatNumber } from "../../../../common";
 import styles from "./NumberInput.module.scss";
 
 function parseNumber(str: string): number {
