+++ components/common/index.ts
@@ -3,3 +3,4 @@ export * from "./articles-inputs";
 export * from "./card"
 export * from "./expandable-panels";
 export * from "./tooltips";
+export * from "./utils";
