+++ components/common/utils/tests/format-number.test.js
