+++ redux/actions/simulations/simulate-cas-types.ts
@@ -1,5 +1,5 @@
-import request from "../../../components/utils/request";
-import { transformCasTypesToData } from "../../../components/utils/transform-cas-types-to-data";
+import request from "../../../components/common/utils/request";
+import { transformCasTypesToData } from "../../../components/common/utils/transform-cas-types-to-data";
 // eslint-disable-next-line no-unused-vars
 import { RootState } from "../../reducers";
 import { formatReforme } from "../format-reforme";
